Added class BombsHandler and Constants file
