+++ .gitignore
@@ -1,4 +1,5 @@
 .idea
+__pycache__/
 /__pycache__/
 *.pyc
 *.pyo
