+++ Constants.py
@@ -1,0 +1,11 @@
+BOMB_WIDTH = 60
+BOMB_HEIGHT = 60
+FIRST_BOMB_STATE = 40
+SECOND_BOMB_STATE = 80
+THIRD_BOMB_STATE = 120
+FOURTH_BOMB_STATE = 160
+
+BOMB_IMAGE_1 = 'assets/bomb/bomb1.png'
+BOMB_IMAGE_2 = 'assets/bomb/bomb2.png'
+BOMB_IMAGE_3 = 'assets/bomb/bomb3.png'
+FIRE_IMAGE = 'assets/bomb/fire.png'
